Add libibmad
libibmad provides low layer IB functions for use by the IB diagnostic
and management programs.
In CentOS 6 and 7 it was not part yet of the rdma-core package. It is
however usually part of a OFED installation and thus should also be
added on conda-forge.
The license was imported from libibmad-1.3.13-1.el7.x86_64.rpm at
/usr/share/doc/libibmad-1.3.13/COPYING.