Add opensm
opensm provides low layer IB functions to interact with the IB subnet
manager.
In CentOS 6 and 7 it was not part yet of the rdma-core package. It is
however usually part of a OFED installation and thus should also be
added on conda-forge.
The license was imported from opensm-3.3.17-1.el6.x86_64.rpm at
/usr/share/doc/opensm-3.3.17/COPYING.