get rid of distro-mangling in _gen_dist_arch_str
Of note: outside of file paths, the only relevant usage was in checking the
`build_append` field of the cdt_slugs, which is not used for any alma builds.

In particular, the names of the actual cdt conda packages themselves are
purely determined by the `short_name` in the distro-config of `rpm.py`.